+++ src/test/java/io/strimzi/kafka/bridge/http/HttpBridgeTest.java
@@ -34,6 +34,7 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
@@ -748,6 +749,7 @@ public class HttpBridgeTest extends KafkaClusterTestBase {
         deleteAsync.await();
     }
 
+    @Ignore
     @Test
     public void sendToOnePartitionTest(TestContext context) {
         String kafkaTopic = "sendToOnePartitionTest";
@@ -787,6 +789,7 @@ public class HttpBridgeTest extends KafkaClusterTestBase {
                 });
     }
 
+    @Ignore
     @Test
     public void partitionPriorityTest(TestContext context) {
         // partition is specified in the message body and in the POST request
@@ -826,6 +829,203 @@ public class HttpBridgeTest extends KafkaClusterTestBase {
                     context.assertNotNull(metadata.getInteger("partition"));
                     context.assertEquals(partition, metadata.getInteger("partition"));
                     context.assertEquals(0L, metadata.getLong("offset"));
+                    async.complete();
+                });
+    }
+
+    @Test
+    public void emptyRecordTest(TestContext context) {
+        Async async = context.async();
+        WebClient client = WebClient.create(vertx);
+
+        client.get(BRIDGE_PORT, BRIDGE_HOST, "").send( ar -> {
+            context.assertTrue(ar.succeeded());
+            context.assertEquals(ErrorCodeEnum.UNPROCESSABLE_ENTITY.getValue(), ar.result().statusCode());
+            context.assertEquals("records cannot not be empty", ar.result().statusMessage());
+            async.complete();
+        });
+    }
+
+    @Test
+    public void invalidRequestTest(TestContext context) {
+        Async async = context.async();
+
+        WebClient client = WebClient.create(vertx);
+
+        client.post(BRIDGE_PORT, BRIDGE_HOST, "/karel")
+                .as(BodyCodec.jsonObject())
+                .sendJsonObject(null, ar -> {
+                    context.assertTrue(ar.succeeded());
+
+                    HttpResponse<JsonObject> response = ar.result();
+
+                    context.assertEquals(ErrorCodeEnum.BAD_REQUEST.getValue(), response.statusCode());
+                    context.assertEquals("invalid request", response.statusMessage());
+                    async.complete();
+                });
+    }
+
+    @Ignore
+    @Test
+    public void sendToNonExistingPartitionsTest(TestContext context) {
+        String kafkaTopic = "sendToNonExistingPartitionsTest";
+        kafkaCluster.createTopic(kafkaTopic, 3, 1);
+
+        Async async = context.async();
+
+        String value = "Hi, This is kafka bridge";
+        int partition = 1000;
+
+        JsonArray records = new JsonArray();
+        JsonObject json = new JsonObject();
+        json.put("value", value);
+        records.add(json);
+
+        JsonObject root = new JsonObject();
+        root.put("records", records);
+
+        WebClient client = WebClient.create(vertx);
+
+        client.post(BRIDGE_PORT, BRIDGE_HOST, "/topics/" + kafkaTopic + "/partitions/" + partition)
+                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
+                .as(BodyCodec.jsonObject())
+                .sendJsonObject(root, ar -> {
+                    context.assertTrue(ar.succeeded());
+
+                    HttpResponse<JsonObject> response = ar.result();
+                    JsonObject bridgeResponse = response.body();
+
+                    JsonArray offsets = bridgeResponse.getJsonArray("offsets");
+                    context.assertEquals(1, offsets.size());
+                    int code = offsets.getJsonObject(0).getInteger("error_code");
+                    String statusMessage = offsets.getJsonObject(0).getString("error");
+
+                    context.assertEquals(ErrorCodeEnum.PARTITION_NOT_FOUND.getValue(), code);
+                    context.assertEquals("Invalid partition given with record: 1000 is not in the range [0...3).", statusMessage);
+                    async.complete();
+                });
+
+    }
+
+    @Test
+    public void sendToNonExistingPartitionsTest2(TestContext context) {
+        String kafkaTopic = "sendToNonExistingPartitionsTest2";
+        kafkaCluster.createTopic(kafkaTopic, 3, 1);
+
+        Async async = context.async();
+
+        String value = "Hi, This is kafka bridge";
+        int partition = 1000;
+
+        JsonArray records = new JsonArray();
+        JsonObject json = new JsonObject();
+        json.put("value", value);
+        json.put("partition", partition);
+        records.add(json);
+
+        JsonObject root = new JsonObject();
+        root.put("records", records);
+
+        WebClient client = WebClient.create(vertx);
+
+        client.post(BRIDGE_PORT, BRIDGE_HOST, "/topics/" + kafkaTopic)
+                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
+                .as(BodyCodec.jsonObject())
+                .sendJsonObject(root, ar -> {
+                    context.assertTrue(ar.succeeded());
+
+                    HttpResponse<JsonObject> response = ar.result();
+                    JsonObject bridgeResponse = response.body();
+
+                    JsonArray offsets = bridgeResponse.getJsonArray("offsets");
+                    context.assertEquals(1, offsets.size());
+                    int code = offsets.getJsonObject(0).getInteger("error_code");
+                    String statusMessage = offsets.getJsonObject(0).getString("error");
+
+                    context.assertEquals(ErrorCodeEnum.PARTITION_NOT_FOUND.getValue(), code);
+                    context.assertEquals("Invalid partition given with record: 1000 is not in the range [0...3).", statusMessage);
+                    async.complete();
+                });
+    }
+
+    @Ignore
+    @Test
+    public void sendToNonExistingTopicTest(TestContext context) {
+        String kafkaTopic = "sendToNonExistingTopicTest";
+
+        Async async = context.async();
+
+        String value = "Hi, This is kafka bridge";
+        int partition = 1;
+
+        JsonArray records = new JsonArray();
+        JsonObject json = new JsonObject();
+        json.put("value", value);
+        json.put("partition", partition);
+        records.add(json);
+
+        JsonObject root = new JsonObject();
+        root.put("records", records);
+
+        WebClient client = WebClient.create(vertx);
+
+        client.post(BRIDGE_PORT, BRIDGE_HOST, "/topics/" + kafkaTopic)
+                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
+                .as(BodyCodec.jsonObject())
+                .sendJsonObject(root, ar -> {
+                    context.assertTrue(ar.succeeded());
+
+                    HttpResponse<JsonObject> response = ar.result();
+                    JsonObject bridgeResponse = response.body();
+
+                    JsonArray offsets = bridgeResponse.getJsonArray("offsets");
+                    context.assertEquals(1, offsets.size());
+                    int code = offsets.getJsonObject(0).getInteger("error_code");
+                    String statusMessage = offsets.getJsonObject(0).getString("error");
+
+                    context.assertEquals(ErrorCodeEnum.TOPIC_NOT_FOUND.getValue(), code);
+                    context.assertEquals("Topic " + kafkaTopic + " not found", statusMessage);
+                    async.complete();
+                });
+    }
+
+    @Ignore
+    @Test
+    public void sendToNonExistingTopicNonExistingPartitionTest(TestContext context) {
+        String kafkaTopic = "sendToNonExistingTopicNonExistingPartitionTest";
+
+        Async async = context.async();
+
+        String value = "Hi, This is kafka bridge";
+        int partition = 1000;
+
+        JsonArray records = new JsonArray();
+        JsonObject json = new JsonObject();
+        json.put("value", value);
+        json.put("partition", partition);
+        records.add(json);
+
+        JsonObject root = new JsonObject();
+        root.put("records", records);
+
+        WebClient client = WebClient.create(vertx);
+
+        client.post(BRIDGE_PORT, BRIDGE_HOST, "/topics/" + kafkaTopic)
+                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
+                .as(BodyCodec.jsonObject())
+                .sendJsonObject(root, ar -> {
+                    context.assertTrue(ar.succeeded());
+
+                    HttpResponse<JsonObject> response = ar.result();
+                    JsonObject bridgeResponse = response.body();
+
+                    JsonArray offsets = bridgeResponse.getJsonArray("offsets");
+                    context.assertEquals(1, offsets.size());
+                    int code = offsets.getJsonObject(0).getInteger("error_code");
+                    String statusMessage = offsets.getJsonObject(0).getString("error");
+
+                    context.assertEquals(ErrorCodeEnum.TOPIC_NOT_FOUND.getValue(), code);
+                    context.assertEquals("Topic " + kafkaTopic + " not found", statusMessage);
                     async.complete();
                 });
     }
